Release/4.3.0 (#2)
* Fixed policyChecker bug

Fix for the policyChecker that will never work as the pathname will never contain query parameters.

* Wrapped jws with try catch

* stringify optionsToValidate when logging

* Limit number of packaged files

* Update README.md

* Update README.md

* 4.2.1

* Update CHANGELOG.md

* Update README.md

Fix for Issue #459
Just spelling / type-o in README.md

* Strip nonbreaking spaces and fix a few lint errors

* Fix test that had whitespace issues

* Version that supports SAML

Pushing information to Read.Me about the version that supports SAML and WSFED, for customer awareness.

* pin version of cache-manager

* fixing the SO tag here.

MSAL was not really a product when passport.js was created, so made sense to have ADAL at that point as the SO tag. but we now should be using MSAL. Updating here so customers get answers faster.

* fix: crash if token is blank

* #341 add in basic proxy agent

rebased and improved readability

* documentation

* fix bad merge

* Added usage comment

* Update README.md

Remove references to waffle.io (as that has been dead for a long time)

* update link in readme

* README: recommend regenerating session by default

The Passport library by default does not prevent against session fixation. Since
the default experience for `passport-azure-ad` is to use cookies, I thought it
would be helpful for the README example to clearly regenerate the session.

This is not a vulnerability with the library itself, but rather just extra caution
common to the passport ecosystem per https://github.com/jaredhanson/passport/issues/192.

* Update lib/bearerstrategy.js

* Release 4.3.0

* Set 4.3.0 to beta.0

* Update bunyan

* Update deps to resolve audit

* 4.3.0-beta.1

* 4.3.0

Co-authored-by: Thomas Haugland Rudfoss <thomas.rudfoss@gmail.com>
Co-authored-by: Mattheu <mattheu@Mattheus-MacBook-Pro.local>
Co-authored-by: Andy Brown <asbrown002@gmail.com>
Co-authored-by: Piotr Kuczynski <piotr.kuczynski@gmail.com>
Co-authored-by: Daryl Thayil <Daryl.Thayil@gmail.com>
Co-authored-by: Nathan Brown <nathan.brown@multiparts.net>
Co-authored-by: Ben Burhans <benburhans@gmail.com>
Co-authored-by: jennyf19 <jeferrie@microsoft.com>
Co-authored-by: Bárbara Castilho <bacast@microsoft.com>
Co-authored-by: Paul Leathwood <paul.leathwood@sky.uk>
Co-authored-by: Jean-Marc Prieur <jmprieur@microsoft.com>
Co-authored-by: jthorel <jthorel@kth.se>
Co-authored-by: hsimah <hamishblake+github@gmail.com>
Co-authored-by: simon.coleman <simon.coleman@hospedia.com>
Co-authored-by: cjancsar <51799534+cjancsar@users.noreply.github.com>
Co-authored-by: Jason Nutter <janutter@microsoft.com>
Co-authored-by: Jeff Wilcox <jeffwilcox@users.noreply.github.com>
Co-authored-by: Jason Nutter <jasonnutter@outlook.com>
Co-authored-by: Ben Bakhar <bebakhar@microsoft.com>